feat(Create NFT): specify startTime/endTime
