+++ src/components/[guild]/RolePlatforms/components/AddRoleRewardModal/components/AddContractCallPanel/AddContractCallPanel.tsx
@@ -3,9 +3,10 @@ import CreateNftForm from "./components/CreateNftForm"
 
 const AddContractCallPanel = ({ onAdd }: AddRewardPanelProps) => (
   <CreateNftForm
-    onSuccess={(guildPlatform) =>
+    onSuccess={(reward) =>
       onAdd({
-        guildPlatform,
+        guildPlatform: reward.guildPlatform,
+        ...reward.rolePlatform,
         isNew: true,
       })
     }
