+++ src/components/[guild]/RolePlatforms/components/EditRewardAvailabilityModal/index.ts
@@ -1,0 +1,6 @@
+import EditRewardAvailabilityModal, {
+  RolePlatformAvailabilityForm,
+} from "./EditRewardAvailabilityModal"
+
+export default EditRewardAvailabilityModal
+export type { RolePlatformAvailabilityForm }
