+++ src/platforms/components/ContractCallPreview.tsx
@@ -1,6 +1,10 @@
-import { Text, Wrap } from "@chakra-ui/react"
+import { Wrap } from "@chakra-ui/react"
 import useSetRoleImageAndNameFromPlatformData from "components/[guild]/AddRewardButton/hooks/useSetRoleImageAndNameFromPlatformData"
-import { CapacityTag } from "components/[guild]/RolePlatforms/components/PlatformCard/components/AvailabilityTags"
+import {
+  CapacityTag,
+  EndTimeTag,
+  StartTimeTag,
+} from "components/[guild]/RolePlatforms/components/PlatformCard/components/AvailabilityTags"
 import useNftDetails from "components/[guild]/collect/hooks/useNftDetails"
 import { useWatch } from "react-hook-form"
 import RewardPreview from "./RewardPreview"
@@ -15,9 +19,17 @@ const ContractCallPreview = (): JSX.Element => {
   const imageUrl = useWatch({
     name: "rolePlatforms.0.guildPlatform.platformGuildData.imageUrl",
   })
+
   const { name, isLoading, maxSupply } = useNftDetails(chain, contractAddress)
 
   useSetRoleImageAndNameFromPlatformData(imageUrl, name)
+
+  const startTime = useWatch({
+    name: "rolePlatforms.0.startTime",
+  })
+  const endTime = useWatch({
+    name: "rolePlatforms.0.endTime",
+  })
 
   return (
     <RewardPreview
@@ -28,9 +40,9 @@ const ContractCallPreview = (): JSX.Element => {
     >
       <Wrap>
         {!!maxSupply && <CapacityTag capacity={Number(maxSupply)} />}
-        <Text colorScheme="gray" fontSize="sm" fontWeight="normal">
-          You'll be able to set up claim start/end date later
-        </Text>
+        {!!startTime && <StartTimeTag startTime={startTime} />}
+
+        {!!endTime && <EndTimeTag endTime={endTime} />}
       </Wrap>
     </RewardPreview>
   )
